Finished Game object.
Upgraded to rails 4 (for following tutorial).
Removed attr_accessible from model, moved it to require() and permit() in controller.
Created games index page with list of games and links to specific game pages.
Verified new game form to have a title for sure.
Added edit and delete routes and links.